refactor(query): Split files into according folders
Added component.html to html file names
