+++ projects/decoupling-test-app/src/hslayers-app/hslayers-app.component.ts
@@ -10,7 +10,7 @@ import {Vector as VectorLayer} from 'ol/layer';
 import {HsConfig} from '../../../hslayers/src/config.service';
 import {HsEventBusService} from 'hslayers-ng/src/components/core/event-bus.service';
 import {HsLayoutService} from 'hslayers-ng/src/components/layout/public-api';
-import {HsQueryFeaturePopupComponent} from 'hslayers-ng/src/components/query/feature-popup.component';
+import {HsQueryFeaturePopupComponent} from 'hslayers-ng/src/components/query/feature-popup/feature-popup.component';
 //import {HsQueryComponent} from '../../../hslayers/src/components/query/query.component';
 //import {HsToolbarPanelContainerService} from 'hslayers-ng/src/components/toolbar/toolbar-panel-container.service';
