+++ projects/hslayers/src/components/query/attribute-row/attribute-row.component.html
