+++ projects/hslayers/src/components/query/default-info-panel-body/default-info-panel-body.component.html
