+++ projects/hslayers/src/components/query/default-info-panel-body/default-info-panel-body.component.ts
@@ -1,9 +1,9 @@
 import {Component, OnInit} from '@angular/core';
-import {HsQueryBaseService} from './query-base.service';
+import {HsQueryBaseService} from '../query-base.service';
 
 @Component({
   selector: 'hs-query-default-info-panel-body',
-  templateUrl: './partials/default-info-panel-body.html',
+  templateUrl: './default-info-panel-body.component.html',
 })
 export class HsQueryDefaultInfoPanelBodyComponent implements OnInit {
   featureInfoExpanded: boolean;
