+++ projects/hslayers/src/components/query/feature-list/feature-list.component.html
