+++ projects/hslayers/src/components/query/feature/feature.component.html
