+++ projects/hslayers/src/components/query/public-api.ts
@@ -1,9 +1,9 @@
-export * from './attribute-row.component';
-export * from './default-info-panel-body.component';
+export * from './attribute-row/attribute-row.component';
+export * from './default-info-panel-body/default-info-panel-body.component';
 export * from './feature-common.service';
-export * from './feature-list.component';
-export * from './feature-popup.component';
-export * from './feature.component';
+export * from './feature-list/feature-list.component';
+export * from './feature-popup/feature-popup.component';
+export * from './feature/feature.component';
 export * from './query-base.service';
 export * from './query-vector.service';
 export * from './query-wms.service';
