+++ projects/hslayers/src/components/query/query.component.html
