+++ projects/hslayers/src/components/query/query.module.ts
@@ -2,16 +2,20 @@ import {CUSTOM_ELEMENTS_SCHEMA, NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {FormsModule} from '@angular/forms';
 
-import {HsDownloadModule} from '../../common/download/download.module';
-import {HsPanelHelpersModule} from '../layout/panels/panel-helpers.module';
-import {HsQueryAttributeRowComponent} from './attribute-row.component';
-import {HsQueryComponent} from './query.component';
-import {HsQueryDefaultInfoPanelBodyComponent} from './default-info-panel-body.component';
-import {HsQueryFeatureComponent} from './feature.component';
-import {HsQueryFeatureListComponent} from './feature-list.component';
-import {HsQueryFeaturePopupComponent} from './feature-popup.component';
 import {NgbDropdownModule} from '@ng-bootstrap/ng-bootstrap';
 import {TranslateModule} from '@ngx-translate/core';
+
+import {HsClearLayerComponent} from './feature-popup-layer/layer-widgets/clear-layer.component';
+import {HsDownloadModule} from '../../common/download/download.module';
+import {HsFeaturePopupLayerComponent} from './feature-popup-layer/feature-popup-layer.component';
+import {HsFeaturePopupWidgetBaseComponent} from './feature-popup-widget-base.component';
+import {HsPanelHelpersModule} from '../layout/panels/panel-helpers.module';
+import {HsQueryAttributeRowComponent} from './attribute-row/attribute-row.component';
+import {HsQueryComponent} from './query.component';
+import {HsQueryDefaultInfoPanelBodyComponent} from './default-info-panel-body/default-info-panel-body.component';
+import {HsQueryFeatureComponent} from './feature/feature.component';
+import {HsQueryFeatureListComponent} from './feature-list/feature-list.component';
+import {HsQueryFeaturePopupComponent} from './feature-popup/feature-popup.component';
 
 @NgModule({
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
@@ -22,6 +26,9 @@ import {TranslateModule} from '@ngx-translate/core';
     HsQueryFeatureListComponent,
     HsQueryDefaultInfoPanelBodyComponent,
     HsQueryAttributeRowComponent,
+    HsClearLayerComponent,
+    HsFeaturePopupWidgetBaseComponent,
+    HsFeaturePopupLayerComponent,
   ],
   imports: [
     CommonModule,
@@ -36,6 +43,9 @@ import {TranslateModule} from '@ngx-translate/core';
     HsQueryFeaturePopupComponent,
     HsQueryFeatureComponent,
     HsQueryFeatureListComponent,
+    HsClearLayerComponent,
+    HsFeaturePopupWidgetBaseComponent,
+    HsFeaturePopupLayerComponent,
   ],
   entryComponents: [HsQueryComponent, HsQueryFeaturePopupComponent],
 })
